pyproject.toml
[project]
name = "household-budget"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.0",
    "djangorestframework-stubs>=3.16.9",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "ruff>=0.15.9",
]
prod = [
    "uvicorn>=0.44.0",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"
exclude = ["*/migrations/*.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle 오류
    "F",  # pyflakes
    "I",  # isort (임포트 정렬)
    "DJ", # Django-specific 규칙
    "UP", # pyupgrade (최신 Python 문법 적용)
    "B",  # bugbear (잠재적 버그 검사)
    "Q",  # flake8-quotes (따옴표 스타일 강제)
]

ignore = ["E203"]  # black과 호환성 위해
fixable = ["ALL"]  # 모든 수정 가능한 문제 자동 수정

[tool.ruff.format]
quote-style = "double"  # black과 일치
